Update stories to latest recommendations
