+++ src/editorial/web/components/logo/Logo.stories.tsx
@@ -1,0 +1,77 @@
+import { Logo, LogoProps } from './Logo';
+import type { Story } from '../../../../@types/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../../../lib/story-intents';
+
+export default {
+	title: 'Editorial/src-ed-logo/Logo',
+	component: Logo,
+	args: {
+		logoType: 'standard',
+	},
+	parameters: {
+		backgrounds: {
+			default: 'brandBackground.primary',
+		},
+	},
+};
+
+const Template: Story = (args: LogoProps) => <Logo {...args} />;
+
+// *****************************************************************************
+
+export const Playground = Template.bind({});
+asPlayground(Playground);
+
+// *****************************************************************************
+
+const TripleTemplate: Story = (args: LogoProps) => (
+	<div>
+		<Logo logoType="standard" />
+		<br />
+		<Logo logoType="anniversary" />
+		<br />
+		<Logo logoType="bestWebsite" />
+	</div>
+);
+
+// *****************************************************************************
+
+export const Desktop = TripleTemplate.bind({});
+Desktop.parameters = {
+	viewport: {
+		defaultViewport: 'desktop',
+	},
+};
+asChromaticStory(Desktop);
+
+// *****************************************************************************
+
+export const Tablet = TripleTemplate.bind({});
+Tablet.parameters = {
+	viewport: {
+		defaultViewport: 'tablet',
+	},
+};
+asChromaticStory(Tablet);
+
+// *****************************************************************************
+
+export const MobileMedium = TripleTemplate.bind({});
+MobileMedium.parameters = {
+	viewport: {
+		defaultViewport: 'mobileMedium',
+	},
+};
+asChromaticStory(MobileMedium);
+
+// *****************************************************************************
+
+export const Mobile = TripleTemplate.bind({});
+Mobile.parameters = {
+	viewport: {
+		defaultViewport: 'mobile',
+	},
+};
+asChromaticStory(Mobile);
+
+// *****************************************************************************
