+++ src/editorial/web/components/logo/tsconfig.json
@@ -6,7 +6,7 @@
     "emitDeclarationOnly": true
   },
   "include": ["."],
-  "exclude": ["dist", "stories.tsx"],
+  "exclude": ["dist", "*.stories.tsx"],
   "references": [
     {
       "path": "../../../../core/helpers"
